v03 correccion headers

+++ ft_bzero.c
@@ -3,17 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: naima <naima@student.42.fr>                +#+  +:+       +#+        */
+/*   By: juagomez <juagomez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/23 12:58:23 by naima             #+#    #+#             */
-/*   Updated: 2024/05/23 13:10:00 by naima            ###   ########.fr       */
+/*   Created: 2024/05/23 16:54:15 by juagomez          #+#    #+#             */
+/*   Updated: 2024/05/23 19:07:01 by juagomez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 // rellena con ceros n bytes en string destino --------
-
 /* declare a temporary pointer */
 /* make the temporary pointer equal to *s converted to a char * */
 /* loop on the temporary pointer while we didn't reach n characters */
@@ -36,14 +35,11 @@ void	*ft_bzero(void *dest, unsigned int len)
 	return (dest);
 }
 
-/* #include <stdio.h>
-#include <strings.h> */
-
 /* int main(void)
 {   
     char str[] = "HolaCaracola";
-    size_t n = 3;
-    size_t nNull = 0;
+    unsigned int n = 3;
+    unsigned int nNull = 0;
 
     printf("test 1 bzero %s \n", bzero(str, n));
     printf("test 1 ft_bzero %s \n", ft_bzero(str, n));
@@ -52,4 +48,4 @@ void	*ft_bzero(void *dest, unsigned int len)
     printf("test 2 ft_bzero %s \n", ft_bzero(str, nNull));
 
     return (0);
-}*/
+} */
